virida_static_api_service/requirements.txt
fastapi==0.61.1
uvicorn
pydantic
sqlalchemy
pymysql
cryptography
pyjwt
alembic
requests
numpy
aiohttp